refactor: project restructure
