+++ src/server/router/routes.ts
@@ -1,6 +1,6 @@
 import { ROUTE_PATH_SEPARATOR } from '../../shared/constants/route-path-separator'
-import { RouteSegments } from '../enums/route-segments.enum'
-import { Route } from '../interfaces/route.interface'
+import { RouteSegments } from '../../shared/enums/route-segments.enum'
+import { Route } from '../../shared/interfaces/route.interface'
 import { usersEndpointHandler } from './route-endpoints-handlers/users-endpoint-handler'
 
 export const routes: Route[] = [
