--- src/server/types/http-method-handlers.type.ts
@@ -1,4 +1,0 @@
-import { HTTPMethods } from '../../shared/types/http-methods.type'
-import { MethodHandler } from './method-handler.type'
-
-export type HTTPMethodHandlers = Record<string, Partial<Record<HTTPMethods, MethodHandler>>>
