+++ src/shared/enums/route-segments.enum.ts
