+++ src/shared/interfaces/route-data.interface.ts
@@ -1,6 +1,6 @@
 import { ParsedUrlQuery } from 'querystring'
 
-import { HTTPMethods } from '../../shared/types/http-methods.type'
+import { HTTPMethods } from '../types/http-methods.type'
 
 export interface RouteData {
   path: string
