+++ src/shared/types/http-method-handlers.type.ts
@@ -1,0 +1,4 @@
+import { HTTPMethodHandler } from './http-method-handler.type'
+import { HTTPMethods } from './http-methods.type'
+
+export type HTTPMethodHandlers = Record<string, Partial<Record<HTTPMethods, HTTPMethodHandler>>>
