+++ src/shared/utils/convert-to-route-data.util.ts
@@ -1,6 +1,6 @@
 import { ParsedUrlQuery } from 'querystring'
 
-import { RouteData } from '../../server/interfaces/route-data.interface'
+import { RouteData } from '../interfaces/route-data.interface'
 import { HTTPMethods } from '../types/http-methods.type'
 
 export const convertToRouteData = (method: HTTPMethods, path: string, query: ParsedUrlQuery, body = ''): RouteData => ({
