+++ src/shared/utils/create-simple-response.util.ts
